\name{read}
\alias{read}
\title{Reads tables}
\usage{
  read(fileName, s, xVar, yVar)
}
\arguments{
  \item{fileName}{Name of the text file}

  \item{s}{Determines what the sep parameter in
  read.table() equals}

  \item{xVar}{First variable for data frame}

  \item{yVar}{Second variable for data frame}
}
\description{
  Creates a data frame with desirable data from data tables
  formatted the way they are in links 2 and 3.
}

